Revert "deleted string to bitmap method"
This reverts commit 30fb3d54df4f803527d4072a7bd853810b2ad5f6.
